fix: restore broken subscription_vault build and re-export needed APIs (#964)
* fix: restore broken subscription_vault build and re-export needed APIs

The fork's main was in a broken, uncompilable state: ~485 lines of type
definitions were missing from types.rs, duplicate use-blocks were left by a
botched merge, several DataKey/Error variants and discriminants were
inconsistent with the frozen layout test, and numerous tests/benches used
pre-SDK-22 call signatures.

This restores the missing definitions and layouts, dedupes imports, makes
internal helpers (calculate_quorum, extend_subscription_ttl,
is_known_instance_discriminant) re-exported, adds missing contractimpl
entrypoints, and updates call sites to the current API across the crate.
The workspace now compiles cleanly and the test suite builds and runs,
with the grace-period clock implementation intact.


* fix: widen ledger sequence to u64 in charge anti-frontrunning salt

env.ledger().sequence() returns u32 in soroban-sdk 22, but the salt buffer
reserves 8 bytes for it, so copy_from_slice panicked with a 4-vs-8 byte
mismatch on every interval charge. Widen to u64 to fix the host panic that
broke auto-pause, grace buyout, fee routing, scheduled cancel, split billing
and subscription-transfer charges.


---------

+++ contracts/subscription_vault/src/charge_core.rs
@@ -471,7 +471,7 @@ pub fn charge_one(
         let mut salt_buf = [0u8; 20];
         salt_buf[..4].copy_from_slice(&subscription_id.to_be_bytes());
         salt_buf[4..12].copy_from_slice(&sub.last_payment_timestamp.to_be_bytes());
-        salt_buf[12..20].copy_from_slice(&seq.to_be_bytes());
+        salt_buf[12..20].copy_from_slice(&(seq as u64).to_be_bytes());
         let salt_input = soroban_sdk::Bytes::from_slice(env, &salt_buf);
         let hash: soroban_sdk::BytesN<32> = env.crypto().sha256(&salt_input).into();
         hash
